The SharpDX commit b9f7519 causes the .NET Native compiler to crash.
This change works around the .NET Native bug by casting to (IntPtr).

I also noticed that DataBuffer had the same issue as DataStream which was
fixed by b9f7519 so I applied the same fix to DataBuffer.

FWIW the code in question calls Interop.Fixed which throws
NotImplementedException.

+++ Source/SharpDX/DataStream.cs
@@ -115,7 +115,8 @@ namespace SharpDX
                 }
                 else
                 {
-                    stream = new DataStream(indexOffset + (byte*)Interop.Fixed(userBuffer), sizeOfBuffer - indexOffset, canRead, canWrite, true);
+                    // The .NET Native compiler crashes if '(IntPtr)' is removed.
+                    stream = new DataStream(indexOffset + (byte*)(IntPtr)Interop.Fixed(userBuffer), sizeOfBuffer - indexOffset, canRead, canWrite, true);
                 }
 
                 return stream;
